Create logistic Regression
This project explores the use of Logistic Regression for performing binary classification to predict whether an individual has diabetes or not. As part of my AI/ML curriculum, this task introduced me to real-world medical data handling, classification concepts, and model performance evaluation.

Using the scikit-learn diabetes dataset (originally designed for regression), I applied a threshold to convert the output into a binary classification problem. The model was trained using Logistic Regression, which is ideal for classifying two categories and works based on the sigmoid function.